missions  Wrapper inlining: local_N = param_N; (debug-mode param copies), (int)this_ptr → (intptr_t)this_ptr, initializing extraout_EDX/extraout_var/extraout_ST0

+++ src/transformed/Horizon/Horizon1/Mission.c
@@ -5,7 +5,7 @@
 #include <time.h>
 
 /* Global variables */
-int this_ptr = 0;
+intptr_t this_ptr = 0;
 char DAT_1002b040[] = "cut";
 char DAT_1002b044[] = "fang";
 char DAT_1002b074[] = "raid";
